Isolated tests depending on PHP 7.1 from other test cases.
